refactor + added totalCost

+++ go/PostOffice/models/label.go
@@ -3,5 +3,5 @@ package models
 type Label struct {
 	ID        string
 	Package   Package
-	TotalCost int32
+	TotalCost float32
 }
